Add the following line below to this page http://www.mendomassage.com/prices.htm  Can you put it at the bottom of the page above the line about special prices.
'We ask $10 per therapist for out-calls to cover time and expense for travel. Thank you.’

Please also remove the Healing Stone, Acupressure and Reiki options

If you can remove those options, please also remove the other references to those options on the following pages:-

On this page,
http://www.mendomassage.com/index.htm
please remove the last paragraph….from “We offer a deep tissue…..….to Prenatal massage.’

On this page,
http://www.mendomassage.com/massage.htm
please remove
The whole of the Healing Stone, Acupressure and Reiki Massage options.